COTECH-691 extract and reuse useJwpWrapperInit() hook in both wrapper versions
